feat: add columns from api, remove getBoards button (used for tests)
